Remove all old EDA figures
This commit removes all of the EDA figures created up to this point. There was an issue created when saving the plots generated with gridExtra::grid.arrange. When creating plots using 'grid.arrange' you have to assign the panelled plot to an object and pass this object to ggsave otherwise the last panel generated in the grid.arrange call is the only one saved to file